requirements.txt
streamlit
mistralai==0.4.2
faiss-cpu
PyYAML
PyPDF2
apscheduler
requests
python-dotenv
mistral
requests